backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
neo4j==5.17.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Vector Search
qdrant-client==1.7.3

# LLM and Embeddings
openai==1.12.0
tiktoken==0.5.2
langchain==0.1.6
langchain-openai==0.0.5
google-generativeai>=0.3.0

# Data Processing
pydantic==2.6.0
pydantic-settings==2.1.0
numpy==1.26.3
pandas==2.2.0

# HTTP and Async
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0

